Hunter/requirements.txt
Django==1.11.9
djangorestframework==3.7.7
SSLyze==1.3.2
requests==2.18.4
cryptography==2.1.4